# src/main/java/org/example/controllers/UserController.java
package org.example.controllers;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.example.dto.requests.CreateUserRequest;
import org.example.dto.requests.ModifyUserRequest;
import org.example.dto.responses.UserResponse;
import org.example.services.UserService;
import org.springframework.web.bind.annotation.*;

@RestController
@RequiredArgsConstructor
public class UserController {
    private final UserService userService;

    @PostMapping("/users")
    public void createUser(@Valid @RequestBody CreateUserRequest createUserRequest) {
        userService.createUser(createUserRequest);
    }

    @PutMapping("/users/{user_id}/project/{project_id}")
    public void assignToProject(@Valid @PathVariable Long user_id,
                                @Valid @PathVariable Long project_id) {
        userService.assignToProject(user_id, project_id);
    }

    @PutMapping("/users/{id}")
    public void modifyUser(@Valid @PathVariable Long id,
                           @Valid @RequestBody ModifyUserRequest modifyUserRequest) {
        userService.modifyUser(id, modifyUserRequest);
    }

    @GetMapping("/users/{id}")
    public UserResponse getUser(@Valid @PathVariable Long id) {
        return userService.getUserById(id);
    }

    @DeleteMapping("/users/{id}")
    public void deleteUser(@Valid @PathVariable Long id) {
        userService.deleteUserById(id);
    }
}